Moving tests into tests dir as per general standard
