+++ tests/fire_test.py
@@ -22,12 +22,11 @@ import os
 import sys
 import unittest
 
-import fire
-from fire import test_components as tc
-from fire import testutils
-
 import mock
 import six
+
+import fire
+from tests import testutils, test_components as tc
 
 
 class FireTest(testutils.BaseTestCase):
