+++ tests/interact_test.py
@@ -18,10 +18,10 @@ from __future__ import absolute_import
 from __future__ import division
 from __future__ import print_function
 
-from fire import interact
-from fire import testutils
-
 import mock
+
+from fire import interact
+from tests import testutils
 
 
 class InteractTest(testutils.BaseTestCase):
