+++ tests/parser_test.py
@@ -19,7 +19,7 @@ from __future__ import division
 from __future__ import print_function
 
 from fire import parser
-from fire import testutils
+from tests import testutils
 
 
 class ParserTest(testutils.BaseTestCase):
