+++ tests/test_components_py3.py
