+++ tests/test_components_test.py
@@ -18,8 +18,7 @@ from __future__ import absolute_import
 from __future__ import division
 from __future__ import print_function
 
-from fire import test_components as tc
-from fire import testutils
+from tests import testutils, test_components as tc
 
 
 class TestComponentsTest(testutils.BaseTestCase):
