+++ tests/trace_test.py
@@ -18,8 +18,8 @@ from __future__ import absolute_import
 from __future__ import division
 from __future__ import print_function
 
-from fire import testutils
 from fire import trace
+from tests import testutils
 
 
 class FireTraceTest(testutils.BaseTestCase):
